refactor: changed charts/ to chart_modals/
