+++ src/components/ribbon/chart_modals/StepModal.tsx
